[libclc] Convert tabs to spaces in CMake
This mirrors a change made to upstream LLVM in
https://github.com/llvm/llvm-project/pull/85634, but since we've made
downstream changes we have to do the work in two places.

Some of this will cause merge conflicts when pulling down upstream LLVM
but I don't know how we can avoid that.

Some of the other files changed in the upstream PR have been left alone
here since we'll soon pull those in anyway.