[Driver][SYCL] Address issue with improper bundler call with -fsycl-link (#13002)
The initial change to modify -fsycl-link for AOT introduced a
modification to the general toolchain flow that is more aligned with the
host link step. This caused a disconnect with the unbundling step from
objects where we were unbundling for host and target, but not providing
the needed output files.

Fix this up to allow for both the host and target output files to be
available. There is an additional fix that will be needed to restrict
this (i.e. only unbundle the target, and not the host), but functionally
what we have in place takes care of the immediate problem at hand.