[SYCL] Flip default printf implementation (#13055)
Currently, the implementation of printf uses a variadic implementation
by default. This proved to be problematic as the implementation would
promote float arguments to doubles, implicitly requring fp64. As a
result of this, an alternative implementation was introduced, but was
made enableable using a __SYCL_USE_NON_VARIADIC_SPIRV_OCL_PRINTF__
directive to avoid problems with targets that did not support it. We
expect most relevant backends to support it now, so we flip the default.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>