[SYCL] Refactor more builtin tests + workaround a bug it uncovered (#12877)
Unfortunately, can't do the same for `sycl::vec` tests because swizzles
can't be captured and a macro seems unavoidable. Also, this includes a
reworked version of https://github.com/intel/llvm/pull/12875 that has
been reverted in https://github.com/intel/llvm/pull/12887.

That also uncovered a bug in our `sycl::length` host implementation when
the project is compiled using GCC. I'm not sure if that's a GCC bug or
yet another instance of UB in `sycl::vec`/`sycl::half` implementation,
but the only option at the moment is to workaround, which I'm doing
here.