[SYCL][AMDGCN] Provide a more helpful --offload-arch error
If multiple SYCL targets are provided, asking users to specify
'-Xsycl-target-backend --offload-arch' will fail for another reason,
which isn't a nice user experience.

Instead, we can conditionally tweak the error message so it shows the
most concise invocation required for it to work every time. This
preserves the old form when only one SYCL target is used - mostly for
consistency - but provides the more verbose form when it's required.