[SYCL][libclc] Add generic addrspace overloads of math builtins (#13015)
The generic implementations of the math builtins which take pointer
arguments were using unqualified address spaces. This could either
resolve to the generic address space or the private address space,
depending on whether the target supports the generic address space or
not.

The newer unified OpenCL C specification is clearer in mandating that
all targets must provide overloads on the explicitly qualified 'private'
address space, as well as optionally defining ones on the (unqualified)
generic address space. This meant that most of these math builtins were
lacking one overload: either the private or generic one, depending on
which target was compiling the builtins.

One notable exception here is NVIDIA, which maps the private and generic
address spaces to the same target address space. Thus declaring builtins
overloaded on these two address spaces results in a mangling clash,
which we can't have. Therefore we now say that NVIDIA targets don't
support the generic address space for the purposes of these builtins. In
reality, the builtins with the private address space are functionally
equivalent to the generic ones, so users won't notice.

For the sake of code clarity, although the 'generic' keyword is
technically reserved, we know that clang defines it to be the
corresponding unqualified generic address space, so we use that to be
explicit. We always compile with clang so it shouldn't be a problem with
portability.

With this we can also enable a LIT test for HIP, which was previously
failing as it couldn't find the generic address space overloads of the
fract and lgamma_r builtins.

There are other builtins that this treatment (may) need applied to, such
as the vload and vstore variants. Those will be handled in a subsequent
patch.