Preserve the dimensionality of ranges
In order to get better performance it is beneficial to preserve
the dimensionality of the range. This new experimental range
rounding does rounding in each dimension.

The runtime can take suggestions on the size of the workgroup in all
directions using SYCL_PARALLEL_FOR_RANGE_ROUNDING_PARAMS=x:y:z. In
the case where -fsycl-exp-range-rounding is used, the middle param is
the workgroup size in all dimensions, so for a 1d range kernel, the
global range will divide {y}. In 2d the global range will be some
number of workgroups of size {y, y}. The same for 3d with {y, y, y}.