[L0] fix a deadlock on a recursive event rwlock
L0, when creating a list of events to wait on, in some cases
was first grabbing a lock on a potentially completed event,
and then tried to get a command list, which sometimes needs to
cleanup all completed events. This caused a deadlock.

This patch moves getting a command list to before the event lock.
But because the lock is required to decide whether this command
list actually needed, we might be wasting time here.