[clang][FE][Cuda] Fix a sm90a cuda arch define check in TargetInfo (#12885)
The original commit from upstream:
[llvm-project/commit/631c6e8](https://github.com/llvm/llvm-project/commit/631c6e834cb07b2769e2c8f1e186dd3a3e0777a1)
conditionally defines `__CUDA_ARCH_FEAT_SM90_ALL` separately from the
CUDA_ARCH, but we break this by making it a either or decision in a
if-else block. Hence, we were not setting correctly the definitions for
upstream's clang -x cuda execution mode for sm90a.

I believe this slipped in as a wrongly resolved merge during an upstream
pulldown.