[CI] Enable post commit for sycl-web pushes
In we may get regressions almost every pulldown, we currently only
expose them when we create the pulldown PR. The pulldown may be delayed
if we get multiple failures.
This is to enable post commit CI to pushes to sycl-web so that we can
know the regressions ealier and may fix the regressions earlier to avoid
delay to pulldown.