[SYCL] No nullptr program_impl (#13155)
The `kernel_impl` has been storing a `program_impl` pointer. But none of
the users of the `kernel_impl` need that, they just need the `PIProgram`
. Conversely, some of the callers constructing `kernel_impl` don't have
a `program_impl` ptr and so we were using nullptr, which leads to
crashes in some simple use cases.

Replacing the `program_impl` ptr with a `PIProgram` member var instead
and ensuring that all callers provide one, with the least amount of
change to the `kernel_impl` interface. Simple test added as well - it
crashes without this PR.