[SYCL][libclc] Add generic addrspace overloads of vload/vstore builtins (#13092)
These definitions and declarations were missing, which would be
problematic for a target which supports the generic address space where
that address space is uniquely mangled. As with the maths builtins, the
NVIDIA target wouldn't see this problem as the 'generic' address space
is mangled identically to the 'private' one, so we were serendipitously
providing the builtins through these.

I couldn't seem to make SYCL actually generate these builtins, but there
is a path from SPIR-V to them (through the OpenCL SPIR-V environment),
so providing them is prudent irregardless. In the absence of tests I
manually verified that there are now as many 'AS0' overloads of these
builtins as there 'AS5'/'AS1'/etc overloads.