switch task sequence enum to new available enum (#2461)
Task Sequence Specification was created without explicitly reserving the 6411 token enum. During spec review, that number was reserved by CooperativeMatrixPrefetchINTEL, but Task Sequence was not updated to use a new enum at that point in time, resulting in development with the old enum. 6199 is the newly (confirmed) reserved token enum.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/08384971f4021b2