[SYCL] Adjust GCC workaround and its scope (#13144)
Previously we were hard-coding an -O2 optimization level for the
'signbit' builtin for all versions of GCC.

Despite this workaround, I found locally that I was unable to build with
GCC versions 12.2, 12.3, and 13.2. Reducing the optimization level to
-O1 allowed me to progress. This seems to follow the bug report already
linked, which had test cases at -O2 which were also failing.

With this in mind, we can also restrict the GCC versions we apply the
workaround to, so that more modern compilers should "just work" without
us having to do anything. That should save someone having to investigate
a performance report a year or so down the line...