[Driver][SYCL] Cleanup redundant dependency and preproc steps
To try and alleviate improper dependency information generated when
-fsycl is enabled, an additionl dependency compilation step was created.
This additional step, while creating the needed information causes
issues when we use -save-temps.  These issues being assosiated with bad
dependencies for the integration header.

Clean up these pathways to be more in line with a typical 'device' and
'host' only compilation, relying on the dependency information from the
device side.

Additional changes can be made when we can clean up the dependency
generation from the host compilation step when we are building the
temporary file.  Removing the append file step would clear this up as
well.