[SYCL] Fix multi_ptr relational operators customised for nullptr. (#13201)
`multi_ptr` relational operators taking a `std::nullptr_t` are written
in a way that assume it is the lowest possible value
(example
https://github.com/intel/llvm/blob/sycl/sycl/include/sycl/multi_ptr.hpp#L1575).
However the C++ specs states there is no ordering requirement
(https://eel.is/c++draft/expr.rel#4.3).

In practice, this is causing issues in the CUDA and AMDGPU backend. For
instance, in CUDA the nullptr in the local address space is a non `0`
value and `0` in this address space is the root of the allocated local
memory.