[SYCL] Remove all uses of `_be` suffix in testing (#13226)
The `_be` suffix was used in sycl/test to identify support for lots of
different backends. For instance: `hip_be`, `cuda_be` etc.

sycl/test-e2e used the same backend identifiers but without the `_be`
suffix. This has caused some confusion, allowing tests to be added to
sycl/test that do not use the `_be` suffix, meaning they are considered
UNSUPPORTED for all backends.

This patch removes all uses of the `_be` suffix in sycl/test, so that
the backend support is aligned with the keywords used in sycl/test-e2e.
Some tests that were not running for any backends now fail, so they are
marked XFAIL.