[UR] Decouple adapter tags & bump OpenCL 3609afc7
This patch makes use of the recently introduced `fetch_adapter_source()`
utility function to decouple fetching all UR adapter source in the event
that a different repo/tag pair is used. To facilitate this
`fetch_adapter_source()` was updated to only perform a fetch if the
adapter name is present in the `SYCL_ENABLE_PLUGINS` variable and the
repo/tag pair does not match the values set in the `UNFIED_RUNTIME_REPO`
and `UNIFIED_RUNTIME_TAG` variables.

To exercise the newly decoupled adapter source mechanism the OpenCL
adapter is bumped to 3609afc7.