[SYCL] Move SYCL Module Splitting to library. Part 2
Add SYCL Module Splitting as a library. ESIMD splitting is not present
in this patch and will be added in the upcoming patch.
Add particular testing tool sycl-module-split that invokes added functionality.
Copy the following tests from llvm/test/tools/sycl-post-link/device-code-split:
    auto-module-split-1.ll
    auto-module-split-2.ll
    auto-module-split-3.ll
    auto-module-split-func-ptr.ll
    basic-module-split.ll
    complex-indirect-call-chain.ll
    one-kernel-per-module.ll
    per-aspect-split-1.ll
    per-aspect-split-2.ll
    per-aspect-split-3.ll
    per-joint-matrix-1.ll
    per-joint-matrix-2.ll
    per-joint-matrix-mad-1.ll
    per-joint-matrix-mad-2.ll
    per-joint-matrix-mad-4.ll
    per-joint-matrix-mad-5.ll
    per-reqd-sub-group-size-split-1.ll
    per-reqd-sub-group-size-split-2.ll
    per-reqd-wg-size-split-1.ll
    per-reqd-wg-size-split-2.ll
    split-with-kernel-declarations.ll
    vtable.ll

The above tests were adjusted to use sycl-module-split tool.

The following list of tests wasn't copied in this patch:
    per-joint-matrix-3.ll
    per-aspect-split-4.ll
    per-joint-matrix-mad-3.ll
    per-reqd-wg-size-split-3.ll
    split-with-func-ptrs.ll

The above tests are exclude because they don't test module splitting
itself. They will be migrated in the upcoming patch.
split-with-func-ptrs.ll requires ESIMD splitting and will be migrated
with the corresponding patch.