[SYCL][XPTI] Fix dangling pointer warnings
Building with GCC 13.2 was showing thousands of lines of warnings due to
the centrality of this header, with the compiler (rightly?) detecting
that ArgsDataPtr may be dangling.

There shouldn't be any performance impact of this change as the
unique_ptr should be just as cheap to initialize with nullptr as a
regular pointer on the fast path.