Both GetElementPtrConstantExpr and GetElementPtrInst may represent access to a buildin variable (#2487)
This PR is to fix the issue when SPIRV Translator expects now to see exactly GetElementPtrInst in an access chain for a builtin variable and crashes in lib/SPIRV/SPIRVUtil.cpp in replaceUsesOfBuiltinVar() with llvm_unreachable() call. See #2486 for the detailed description. Attached is a test case that demonstrates the problem and leads to a crash in current version(s) of SPIRV Translator.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/a7051670770f471