[libclc] Update build system to reflect LLVM upstream changes
This brings the libclc build system more in line with LLVM upstream,
which was recently changed significantly. It throws out a fair bit of
dead CMake code too.

Upstream, libclc was changed away from using custom languages to built
and towards custom commands. This makes in-tree builds a lot simpler and
less hacky, as well as a few other benefits.

Unfortunately, due to peculiarities involving the `-c` and `-x cl`
flags, and the combination of the two which enable/disable different
optimizations, I was seeing significant changes to the builtin libraries
when using the same flags as upstream. As such, I've had to diverge from
upstream and not use `-x cl` when compiling builtin files. This should
keep the builtin libraries identical to before this change, but it's not
ideal to be compiling ostensibly OpenCL files as if they were another
language.

The intermediate bytecode files and builtins have moved a bit, but the
final libraries (both remangled and non-remangled) are in the same
place.