[SYCL] Handle PI_EVENT_STATUS_QUEUED return value (#13024)
https://github.com/intel/llvm/pull/9944 introduced a hack in the OpenCL
PI plugin which maps PI_EVENT_QUEUED to PI_EVENT_SUBMITTED. This hack is
problematic because it will make SYCL-RT assume that an event has been
flushed when it might not have been. At the moment, this could result in
buggy behaviour on the OpenCL backend when using multiple queues. The
OpenCL specification states:

> "To use event objects that refer to commands enqueued in a
command-queue as event objects to wait on by commands enqueued in a
different command-queue, the application must call a clFlush or any
blocking commands that perform an implicit flush of the command-queue
where the commands that refer to these event objects are enqueued."

This PR, moves the mapping from UR to `get_event_info()` inside sycl-rt
which avoids the issue because it is a separate code path from
`flush_if_needed()`

Corresponding UR PR:
https://github.com/oneapi-src/unified-runtime/pull/1440