Fix Coverity wrapper escape issue
Issue was introduced in #1220 where a new allocation is created in a
`std::unique_ptr` to enable buffer mapping. The `.get()` member function
is called before the `std::unique_ptr` is passed into the `BufferMem`
constructor. However, move semantics were not being explicitly followed.
The `BufferMem` constructor was taking the `std::unique_ptr` by l-value
refernece, then calling `std::move()` on it. This triggered a
wrapper-escape, use-after-free defect in Coverity. This fix is to
explicitly `std::move()` the `std::unique_ptr` into the `BufferMem`
constructor, and also update the `BufferMem` constructor to take an
r-value reference instead of an l-value reference.