[Driver][SYCL] Remove append step from integration footer with clang
When using the clang compiler (current compiler) to perform the host
compilation with offloading, use the -include-footer option to pull in
the footer as opposed to appending the footer using the append-file
tool.  The use of the append-file tool is still required when using an
external host compiler.

This improves the host compilation by working with the proper source
file as opposed to a temporary file.  This impacts dependency
generation, debug information, preprocessing information.