[CI] Redirect and fix email-check workflow (#13375)
https://github.com/intel/llvm/commit/1bed6397a3c47eb3ca19b70f8d15a7860da0792d
resolved a merge conflict in .github/workflows/email-check.yaml by
reintroducing a removed "uses" key. However, this makes for an invalid
configuration. This commit removes the invalid conflict resolution.

Additionally, it changes the check for repository in the workflow to
require intel/llvm. This should make it run as part of our CI.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>