Fix of ParentIdx mismatch for DebugTypeInheritance in SPIRV.debug.h (#2490)
Fixed ParentIdx was mismatched for DebugTypeInheritance type in context of NonSemantic.Shader.DebugInfo.100. That lead to heap buffer overflow in SPIRVExtInst::getExtOp getter when instruction was incorrectly casted
to SPIRVExtInst as parent of the culprit instruction in SPIRVToLLVMDbgTran::getDIBuilder. The missmatch happen because in previous used standard OpenCL.DebugInfo.100 DebugTypeInheritance had Child field as zero indexed argument. In newer standard that field is removed.

Original commit:
https://github.com/KhronosGroup/SPIRV-LLVM-Translator/commit/c6e9fea6d0f70f1