[SYCL][Doc] Remove get_child_group() documentation
get_child_group() was originally designed assuming there would be a fixed
hierarchy of groups, and is therefore incompatible with extensions that would
modify the hierarchy (e.g., "clusters" on NVIDIA GPUs).

This interface should be replaced by a more general and explicit mechanism for
partitioning groups that does not make such assumptions.

Signed-off-by: John Pennycook <john.pennycook@intel.com>