[SYCL] fix persistent cache test (#13466)
On Windows the persistent cache is still working but the path string
when writing is one mishmash of Linux and Win separators and on read a
different mishmash for the same file. Corrects this test on Win.

Also am expanding it to confirm write/reuse/reuse is working on both Lin
and Win, not just Win.