[SYCL][JIT] Load SYCL JIT lazily (#13433)
So far, the SYCL JIT library, used e.g. to implement the kernel fusion
extension, was linked directly into `libsycl`.

This led to a significant increase in library binary size of `libsycl`.
To avoid the corresponding startup overhead, in particular if
applications do not use the kernel fusion extension, this PR implement
lazy loading of the JIT library.

Instead of linking the library into `libsycl`, the JIT library is built
as a standalone shared library (`libsycl-fusion.so`). This library is
only loaded at runtime on the first use of the JIT compiler and
applications not using a feature that requires the JIT will not load the
library at all.

The dynamic loading happens through `dlopen` and `dlsym`, through
existing functionality provided by PI and also used for the SYCL online
compiler and kernel compiler.

This PR also includes necessary changes to the interface of the JIT
library to allow dynamic loading, mainly changing interface entry points
to `extern "C"` and changing the options interface to not require
template functions.

---------

Signed-off-by: Lukas Sommer <lukas.sommer@codeplay.com>