[GHA] Separate out hardware workflows
The goal of this patch is to make a failure of one HW build & test job
only cancel jobs running on the same HW tag rather than all HW jobs for
all adapters. It does this by moving the hardware jobs out of the
`cmake.yml` workflow into the new `build-hw-reusable.yml` workflow,
which as the name suggests is reusable. This reusable workflow is then
used in by adapter specific jobs in the `cmake.yml` workflow to display
jobs in a group which is collapsible.