Use consistent indent style throughout
Until now we've had multiple styles throught the project. This doesn't
buy us anything for contributors as each source file might be different
- totally negating the point of having a clang-format style in the first
place - indeed I started debugging clang-format since it was apparently
not formatting my adapter code as 4 spaces after my editor had loaded
the editconfig file specifying 4 spaces. I wasted an hour on that!

Here we codify everything as LLVM style, which seems to be used by the
majority of the project.

We also bump the clang-format version requirement to v18.1 to align with
dpc++.