[SYCL][E2E] Fix some tests in multi-device mode
The atomic tests can be run by multiple SYCL devices, in which case the
compiler needs to be told which device the `--cuda-gpu-arch` parameter
applies to.

The changes to the bfloat16 tests try and restore earlier behaviour,
where the `aspect-ext_oneapi_bfloat16_math_functions` feature was
preventing other devices from running. It also removes RUN lines from a
shared header file (which we don't consider a valid test file) which may
mislead people into thinking it itself is a runnable test.