Fix adding event to queue cache
in single-device scenario.

When event is an internal event (when using queue with discard
events property) the UrQueue is set to nullptr. This means that
inside addEventToContextCache the event will be added to a
multi-device cache which is wrong. This causes unbounded growth
of the multi-device event cache if there are no multi-device
events requested from the cache.