[SYCL][Graph] Update begin_recording and end_recording (#13480)
- Removes bool return value from begin_recording() and end_recording()
- Adds missing property_list parameter to the implementation of
begin_recording()

The bool return value was removed because there are no good usecases for
it. The sycl way to handle errors is to return exceptions. The original
reasoning to have a bool return value was that the application would
likely ignore the exceptions in this entrypoint. There is no reason to
believe that they will check the bool return either if it serves no
practical purpose.

This is a breaking API change, but as the extension is in experimental
status this is permissible.