[SYCL][ABI-Break] Fold host_half_impl::half into half_impl::half
This commit folds the implementation of host_half_impl::half into
half_impl::half and making the vector element representation the same
as the half representation. This allows us to avoid strict alias
violation for half vectors in their operator[] implementations.

Note that this is marked as an ABI break as it removes symbols on
Windows, despite these symbols never being in the library.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>