stop batching multiple CmdList execution when waiting for barriers
The current implementation of queue command batching appears to
assume that getAvailableCommandList is followed by a single
executeCommandList, and it's not allowed to, for example,
create a list of CmdLists for later execution. And there's
an assert to check for this.

This makes sense, since it doesn't really make sense to batch
things across many different CmdLists. However, the implementation
of urEnqueueEventsWaitWithBarrier was doing exactly that.

This patch disables batching in executeCommandList, fixing
a assert failure during urEnqueueEventsWaitWithBarrier
when many potentially open command lists are in use.