[SYCL] Add work_group_num_dim metadata
Emit metadata to describe number of dimensions specified in
reqd_work_group_size. This is needed in order to be able to use that
metadata correctly, since it was specified for OpenCL, and SYCL
piggy-backs on it, backends correctly assert if all 3 dimensions are not
provided. work_group_num_dim allows the compiler to pad the missing
dimensions with 1, while preserving the notion of how many dimensions
were specified.