[HIP] Add a workaround for crashes on gfx1031
When calling hipMemGetInfo, an internal problem querying the amount of
available global memory causes the function to return
hipErrorInvalidValue. This is indistinguishable from us doing something
wrong, and it (rightly) causes an assertion failure. This is a
particularly egregious issue because we query this property and crash
during the execution of `sycl-ls --verbose`.

This architecture is not officially supported by HIP, so we can't file a
bug and shouldn't expect an official fix for this issue any time soon.
It is possible, however, that we'll need to extend this workaround with
other architectures over time.