[Driver][SYCL] Use LLVM-IR based device libraries for device linking
This is try #2 for this - original change was suspected of breaking the
device library builds.

This change represents 2 changes.
1.  Update the build to create LLVM-IR based device libraries.

This is done by using the existing -fsycl-device-only functionality to
create the device libraries that are used during device link.

2.  Update the compiler driver to use the LLVM-IR based device
    libraries.

When performing the device linking instead of performing a full unbundle
of the device libraries and using the device binary that was extracted.

This streamlines the device linking behaviors, completely removing the
cumbersome unbundling step from the process, freeing up a number of
exteral tool calls.