Update tests
It doesn't make sense to check that a context has been set in the
background when a buffer is created.

Also, since images and buffers are now allocated lazily, we can't know
if the mem/image creation will fail at `urMemImageCreate`. Maybe this is
a problem and we should allocate eagerly for single device contexts.