Improve CTS device code generation
Some improvements to the device code generation cmake:
* Use `clang-offload-extract` instead of running the binary
  directly.
* Set up dependencies correctly so that re-running builds will
  generate missing IR files if they failed to build prior.
* Dynamically generate the list of device_code files - from now on,
  any .cpp file in the device_code directory will automatically be
  treated as device code and generate IR files.

This change drops CTS support for platforms that support multiple
devices with different IR types, however the build system didn't
support that anyway, nor do any of our adapters.

This change requires a `clang-offload-extract` binary to be
available. By default it is assumed to be in the same directory as
the provided DPC++ binary (which should work for most builds),
however the CMake variable `UR_DEVICE_CODE_EXTRACTOR` is provided
to override this.