[Driver][SYCL] Pass full set of sycl-post-link options to linker wrapper
When using the new offloading model, pass all of the options associated
with sycl-post-link to the clang-linker-wrapper.  These options will be
used by the clang-linker-wrapper to setup the sycl-post-link calls
during the device linking step.

This is more of a first step/general application here, as there are
nuances in the options that should be used based on the device objects
themselves.  The options passed are based on what is known by the driver
specifically for that particular invocation.