[NFC] Bump CMake verson to 0.10.0
Bumping to v0.10.0 exposed a bug in the spec generation scripts where
versions were being converted to `float` and then used doing version
comparisons. Converting `0.10` to `float` results in `0.1` which is
an incorrect representation of that version. To address this the scripts
have been updated, including the new `Version` class which handles
version comparisons, to never use `float` to represent versions.