[Driver][SYCL][FPGA] Use bundled device libraries for FPGA targets
The change that introduced the ability to use device libraries as IR
files does not play well with the FPGA target.  The IR generated with
spir64_fpga is different, causing issues as the bitcode based device
libraries are built with spir64.

Update the compilation flow specific to FPGA to use the bundled device
libraries as opposed to the bitcode files.