Only test DEVICE_DEFAULT query when queue is device default.
Some supported CL drivers return an error rather than NULL if you try
this query on a queue that wasn't created with the ON_DEVICE and
DEVICE_DEFAULT flags. This isn't correct according to the CL spec but
limiting our testing to the supported case is still preferable to adding
an exception to the match files.