[SYCL][NATIVECPU][CI] Enable Native CPU without OCK on linux build (#13052)
This PR enables the Native CPU plugin in the build jobs for Linux. It is
the first PR in series, with the goal of enabling full `e2e` testing for
the Native CPU plugin on both Linux and Windows.

This initial PR builds the Native CPU plugin without the oneAPI
Construction Kit. While the oneAPI Construction Kit is currently a key
component for Native CPU, I would prefer to keep things as simple as
possible, and enable CI testing for additional features incrementally,
in order to gradually assess the impact of the Native CPU plugin in the
overall CI workflow.