[SYCL][Graph][CUDA] Workaround Windows shared USM
On a CUDA & Windows setup with **system** USM is used there is an
issue with using the allocations concurrently in both device commands
and host-tasks. This manifests as a page fault in the host-task when
trying to access the USM allocation, presumably due to the contention
between the device and host for the memory.

This appears in SYCL-Graph tests where the graph has multiple roots,
allowing host-tasks branching from one root to run concurrently with
device commands from the other root. I've created a more minimal test
`test-e2e/USM/host-task.cpp` which exhibits the same issues.

This commit works around the failure in graph tests by using host USM
allocations instead of shared allocations.