[SYCL][Driver] Refactor device config parsing to better match HIP and CUDA targets (#13617)
Refactors the device config parsing by making it agnostic of the way the
SYCL device is targeted, either by passing a device triple plus arch
special string to `-fsycl-targets` or a just the device arch to
`--offload-arch` (mostly for CUDA and HIP). This is achieved by also
moving the parsing logic into `SYCLActionBuilder`, in order to make use
of the device architecture bound to the offloading toolchain to lookup
in the DeviceConfig file's TargetTable accordingly.
This PR also introduces HIP and CUDA target info in the
`DeviceConfigFile.td` which is currently not complete but a starting
point to build on.