[SYCL] [Joint Matrix] Modularized tests (#13785)
Some tests had a lot of duplicate code. In many cases functions could be
collapsed to a general `matrix_verify_op()` function that takes in a
lambda, instead of having a unique function for each operation.

Although it may sees that the implementations of `matrix_verify_op()` is
itself is repeated between multiple files that is by design. The
implementations differ by the `use` parameter, which dimensions are used
to traverse the matrix, and packing behaviour of tiles. Although this
too can be abstracted out to template parameters or function arguments,
this will grow already complicated interface to unreasonable sizes.