[Driver][SYCL] Do not pass along implied PVC specific values for AOT
When compiling for AOT GPU, we passed along default device settings for
PVC regardless of intended device.  Update this behavior to only pass
along the PVC specific options to ocloc when the matching device (i.e.
-device pvc) is also supplied.

The device can be specific with the typical device value (pvc) or via
the -fsycl-targets=intel_gpu_pvc option setting or -device values that
match with PVC that match a given version or HEX value.

The default values will continue to be passed along for JIT.