[Unittests] Fix device info query for shared USM
The unittests have a default mock implementation of device info queries
which returns success for any query made. However, for cases where the
query does not have a case in the implementation, nothing is done to the
return values. For support queries this may mean that one call to the
query returns true and another returns false. This has been seen for
shared USM allocation support, which caused flaky failures when
unittests would call malloc_shared.

This commit:
* Changes the default implementation of the device info query to return
true for shared USM allocation support, like it already does for host
and device variants.
* Changes the default implementation of the device info query to fill
the return value with 0 and set the return type size to 1 when the info
query is unrecognized. This means unittests should always use
`redefineAfter` for the device info query.
* Fixes tests that assumes device info queries may be device-dependent
in unittests.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>