[SYCL][Graph][CUDA] Skip unsupported Windows E2E tests (#13764)
On a CUDA & Windows setup when shared USM is used there is an issue with
using the allocations concurrently in both device commands and
host-tasks. This is based on an underlying CUDA restriction:
https://forums.developer.nvidia.com/t/cudamallocmanaged-clarification-needed/67611

> Applications running on Windows (whether in TCC or WDDM mode) or macOS
will use the basic Unified Memory model as on pre-6.x architectures even
when they are running on hardware with compute capability 6.x or
higher.”

> “Simultaneous access to managed memory on devices of compute
capability lower than 6.x is not possible,”

Therefore, simultaneous access to managed memory on Windows is not
possible.

This appears in SYCL-Graph tests where the graph has multiple roots,
allowing host-tasks branching from one root to run concurrently with
device commands from the other root. With the issue manifesting as a
page fault in the host-task when trying to access a USM allocation. I've
created a more minimal test `test-e2e/USM/host-task.cpp` which exhibits
the same issues.